server/pyproject.toml
[tool.poetry]
name = "genericsuite-codegen"
version = "1.0.0"
description = "Generic Suite CodeGen - RAG AI system for GenericSuite documentation and code generation"
authors = ["Carlos J. Ramirez <info@genericsuite.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "genericsuite_codegen"}]

[tool.poetry.dependencies]
python = "^3.12,<3.14"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.5.0"
pydantic-ai = "^1.0.1"
pymongo = "^4.6.0"
openai = "^1.3.0"
sentence-transformers = "^2.2.2"
transformers = "^4.36.0"
torch = "^2.1.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.0"
litellm = "^1.17.0"
httpx = "^0.28.1"
aiofiles = "^23.2.0"
gitpython = "3.1.45"
fastmcp = "^2.12.2"
pypdf = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true